feat: zustand로 유저 정보 관리
