+++ src/app/(auth)/login/page.tsx
@@ -9,39 +9,21 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import FormWrapper from "../_components/form-wrapper";
 import AuthRedirect from "../_components/auth-redirect";
-import { User } from "@/types/user-type";
-
-interface LoginData {
-  user: User;
-  accessToken: string;
-  refreshToken: string;
-}
+import useUserStore from "@/store/user-store";
+import login from "@/api/auth/login";
 
 interface FormData {
   email: string;
   password: string;
 }
 
-const login = async (data: any): Promise<LoginData | undefined> => {
-  try {
-    const response = await instance.post("/auth/signIn", data);
-    console.log(response);
-
-    if (!response) throw new Error("데이터를 불러오는데 실패했습니다.");
-
-    return response.data;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
 const Page = () => {
   const [formData, setFormData] = useState<FormData>({
     email: "",
     password: "",
   });
-  const [user, setUser] = useState<User>();
   const router = useRouter();
+  const { user, setUser, clearUser } = useUserStore((state) => state);
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
     setFormData((prevData) => ({ ...prevData, [e.target.id]: e.target.value }));
@@ -53,11 +35,14 @@ const Page = () => {
     if (!response) throw new Error("로그인에 실패했습니다.");
 
     setUser(response.user);
+    console.log(user);
 
-    sessionStorage.setItem("accessToken", response?.accessToken);
-    localStorage.setItem("refreshToken", response?.refreshToken);
+    // sessionStorage.setItem("accessToken", response?.accessToken);
+    // localStorage.setItem("refreshToken", response?.refreshToken);
+    document.cookie = `accessToken=${response?.accessToken}`;
+    document.cookie = `refreshToken=${response?.refreshToken}`;
 
-    router.push("/");
+    // router.push("/");
   };
 
   // /users/me API 테스트를 위한 함수
@@ -73,6 +58,7 @@ const Page = () => {
   return (
     <FormWrapper>
       {/* // TODO: 로고 링크 추가 */}
+
       <Logo className="mb-10 h-[30px] w-[104px] text-gray-1100" />
       <form action={onSubmit} className="flex-col-center mb-4 gap-[60px]">
         <div className="flex flex-col gap-8">
@@ -111,7 +97,6 @@ const Page = () => {
           url="/signup"
         />
       </div>
-      {/* <Button label="내 정보" onClick={getMe} /> */}
 
       {user && (
         <div>
